move FirebaseConfig.tsx from src/config/ to /config/
