+++ src/services/CollectionSetService.tsx
@@ -1,5 +1,5 @@
 import { collection, getDocs, query, where } from "firebase/firestore";
-import { database } from "../config/FirebaseConfig";
+import { database } from "../../config/FirebaseConfig";
 import { CollectionCard } from "../types/CollectionCard";
 import { CollectionSet } from "../types/CollectionSet";
 import { SetAndCard } from "../types/MergedCollection";
